+++ fPayment_Slip.cs
@@ -81,7 +81,7 @@ namespace DuLich
         private void btn_BookNow_Click(object sender, EventArgs e)
         {
 
-            datphong.TongThanhToan = float.Parse(lblPrice.Text);
+            datphong.TongThanhToan = double.Parse(lblPrice.Text);
             DatPhongDAO datPhongDAO = new DatPhongDAO();
             KhachHangDAO khachHangDAO = new KhachHangDAO();
             datPhongDAO.AddDatPhong(datphong);
